Reorganise destination test code
