+++ products/batch_exports/backend/api/destination_tests/destination_tests.py
@@ -1,0 +1,18 @@
+from products.batch_exports.backend.api.destination_tests.base import DestinationTest
+from products.batch_exports.backend.api.destination_tests.bigquery import BigQueryDestinationTest
+from products.batch_exports.backend.api.destination_tests.s3 import S3DestinationTest
+from products.batch_exports.backend.api.destination_tests.snowflake import SnowflakeDestinationTest
+
+
+def get_destination_test(
+    destination: str,
+) -> DestinationTest:
+    """Resolve a destination to its corresponding `DestinationTest` implementation."""
+    if destination == "S3":
+        return S3DestinationTest()
+    elif destination == "BigQuery":
+        return BigQueryDestinationTest()
+    elif destination == "Snowflake":
+        return SnowflakeDestinationTest()
+    else:
+        raise ValueError(f"Unsupported destination: {destination}")
